requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.6
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.6
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.116.1
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
fonttools==4.59.0
greenlet==3.2.4
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
ipykernel==7.0.1
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.8
llvmlite==0.45.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.5
matplotlib-inline==0.2.1
mdurl==0.1.2
narwhals==2.0.1
nest-asyncio==1.6.0
numba==0.62.1
numpy==2.2.6
optuna==4.5.0
packaging==25.0
pandas==2.3.1
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
plotly==6.2.0
prompt_toolkit==3.0.52
psutil==7.1.2
psycopg2==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.3
pycparser==3.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.33.1
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.42.1
shap==0.49.1
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.47.2
threadpoolctl==3.6.0
tomli==2.3.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
xgboost==3.1.1